Code improvement and cleaning
Cleaned code for the github release.
Also added new encryption system;
token is encrypted/decrypted with AES, the key is currently hardcoded in Program.cs, do what-ever you want with that.
It's also salted, using the MD5 hash of the compiled binary, so whenever you compile/update, the encryption / decryption will stop working and you'll have to put a new token in. Modify this as you want